A visual of Fourier Series
An animation that shows generation of sine waves from a circle and how summing the odd harmonics gradually constructs a square wave. This is an implementation of the Fourier series, that says you can disintegrate any signal into a series of sine and cosine waves or create any signal using the same.